+++ built/grammar/AbstractRuntimeProgram.d.ts
@@ -26,12 +26,12 @@ declare abstract class AbstractRuntimeProgram extends AbstractRuntimeNode {
     getInPlaceSyntaxTreeWithNodeTypes(): string;
     protected _getSyntaxTreeHtml(): string;
     getTreeWithNodeTypes(): string;
-    getWordTypeAtPosition(lineIndex: number, wordIndex: number): any;
+    getCellTypeAtPosition(lineIndex: number, wordIndex: number): any;
     getWordHighlightScopeAtPosition(lineIndex: number, wordIndex: number): types.highlightScope | undefined;
-    private _cache_programWordTypeStringMTime;
+    private _cache_programCellTypeStringMTime;
     private _cache_highlightScopeTree;
     private _cache_typeTree;
-    protected _initWordTypeCache(): any;
+    protected _initCellTypeCache(): any;
     getCompiledProgramName(programPath: any): any;
 }
 export default AbstractRuntimeProgram;